+++ src/lib/statistics/chunk_statistics/histograms/equal_num_elements_histogram.hpp
@@ -48,7 +48,7 @@ class EqualNumElementsHistogram : public AbstractHistogram<T> {
                  const std::shared_ptr<const ValueColumn<int64_t>> count_column, const size_t max_num_buckets) override;
 
   static EqualNumElementsBucketStats<T> _get_bucket_stats(const std::vector<std::pair<T, uint64_t>>& value_counts,
-                                                          const uint64_t max_num_buckets);
+                                                          const size_t max_num_buckets);
 
   BucketID _bucket_for_value(const T value) const override;
   BucketID _lower_bound_for_value(const T value) const override;
